# pkg/app/response.go
package app

import (
	"github.com/gin-gonic/gin"

	"github.com/fish2016/go-gin-dev/pkg/e"
)

type Gin struct {
	C *gin.Context
}

type Response struct {
	Code int         `json:"code"`
	Msg  string      `json:"msg"`
	Data interface{} `json:"data"`
}

func (g *Gin) Response(httpCode, errCode int, data interface{}) {
	g.C.JSON(httpCode, Response{
		Code: httpCode,
		Msg: e.GetMsg(errCode),
		Data: data,
	})
	return
}